Correcting contacts processing functions
