+++ routes/api/contacts.js
@@ -2,32 +2,7 @@ const express = require('express')
 
 const router = express.Router()
 
-const Joi = require('joi');
-
-const schemaAddContact = Joi.object({
-  name: Joi.string()
-    .pattern(/^[a-zA-Zа-яА-Я]+(([' -][a-zA-Zа-яА-Я ])?[a-zA-Zа-яА-Я]*)*$/)
-    .required(),
-
-  email: Joi.string()
-    .email({ minDomainSegments: 2})
-    .required(),
-    
-  phone: Joi.string()
-    .pattern(/\+?\d{1,4}?[-.\s]?\(?\d{1,3}?\)?[-.\s]?\d{1,4}[-.\s]?\d{1,4}[-.\s]?\d{1,9}/)
-    .required(),
-})
-
-const schemaUpdateContact = Joi.object({
-  name: Joi.string()
-    .pattern(/^[a-zA-Zа-яА-Я]+(([' -][a-zA-Zа-яА-Я ])?[a-zA-Zа-яА-Я]*)*$/),
-
-  email: Joi.string()
-    .email({ minDomainSegments: 2}),
-    
-  phone: Joi.string()
-    .pattern(/\+?\d{1,4}?[-.\s]?\(?\d{1,3}?\)?[-.\s]?\d{1,4}[-.\s]?\d{1,4}[-.\s]?\d{1,9}/),
-})
+const { schemaAddContact, schemaUpdateContact } = require('../../helpers/validation_schema')
 
 const { listContacts,
   getContactById,
@@ -50,37 +25,36 @@ router.get('/:contactId', async (req, res, next) => {
 })
 
 router.post('/', async (req, res, next) => {
-  const { name, email, phone } = req.body
-  const result = schemaAddContact.validate({ name: name, email: email, phone: phone })
-  if (!name || !email || !phone || result.error) {
-    res.status(400).json({ "message": "missing required name field" })
+  const { value, error } = schemaAddContact.validate(req.body)
+  if (error) {
+    res.status(422).json({ "message": error.message })
   } else {
-    res.status(201).json(await addContact(req.body))
-  }  
+    res.status(201).json(await addContact(value))
+  }
 })
 
 router.delete('/:contactId', async (req, res, next) => {
   const { contactId } = req.params
   const contacts = await listContacts()
   if (!contacts.find(contact => contact.id === contactId)) {
-    res.status(404).json({"message": "Not found"})
+    res.status(404).json({ "message": "Not found" })
   } else {
     await removeContact(contactId)
-    res.json({"message": "contact deleted"})
+    res.json({ "message": "contact deleted" })
   }
 })
 
 router.put('/:contactId', async (req, res, next) => {
   const { contactId } = req.params
-  const result = schemaUpdateContact.validate({ name: req.body.name, email: req.body.email, phone: req.body.phone })
+  const { value, error } = schemaUpdateContact.validate(req.body)
   const contacts = await listContacts()
   if (!contacts.find(contact => contact.id === contactId)) {
-    res.status(404).json({"message": "Not found"})
-  } else if (!req.body || result.error) {
-    res.status(400).json({"message": "missing fields"})
+    res.status(404).json({ "message": "Not found" })
+  } else if (Object.keys(req.body).length === 0 || error) {
+    res.status(400).json({ "message": "missing fields" })
   } else {
-    res.json(await updateContact(contactId, req.body))
-  }  
+    res.json(await updateContact(contactId, value))
+  }
 })
 
 module.exports = router
